% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ArrayDataCollection.R
\docType{class}
\name{ArrayDataCollection-class}
\alias{ArrayDataCollection-class}
\alias{ArrayDataCollection}
\title{Array Data Collection}
\description{
A class for working with continuous expression data from arrays (antibody arrays,
gene expression microarrays, etc.). Unlike CountDataCollection, this class accepts
continuous numeric data including negative values (e.g., log-transformed data).
}
\details{
Users are expected to provide pre-normalized data. No normalization is performed
by veupathUtils methods.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame of numeric expression values with features as columns and samples as rows}

\item{\code{sampleMetadata}}{A data.frame of metadata about samples with samples as rows and metadata variables as columns}

\item{\code{recordIdColumn}}{The name of the column containing IDs for the samples}

\item{\code{ancestorIdColumns}}{A character vector of column names representing parent entities of the recordIdColumn}

\item{\code{imputeZero}}{A logical indicating whether NA/null values should be replaced with zeros}
}}

