% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Bin.R
\name{as.Date,Bin-method}
\alias{as.Date,Bin-method}
\title{Convert a Bin's start and end values to dates}
\usage{
\S4method{as.Date}{Bin}(x)
}
\arguments{
\item{x}{Bin}
}
\value{
Bin
}
\description{
Given a Bin object with slots binStart and binEnd, convert
the binStart and binEnd to dates
}
