% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Bin.R
\name{as.Date,BinList-method}
\alias{as.Date,BinList-method}
\title{Convert BinList bin starts and ends to dates}
\usage{
\S4method{as.Date}{BinList}(x)
}
\arguments{
\item{x}{BinList}
}
\value{
BinList
}
\description{
Given a BinList object, convert all binStarts and binEnds to dates and return the BinList.
}
