% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Bin.R
\name{as.numeric,Bin-method}
\alias{as.numeric,Bin-method}
\title{Convert a Bin with character start and end values to numeric}
\usage{
\S4method{as.numeric}{Bin}(x)
}
\arguments{
\item{x}{Bin}
}
\value{
Bin
}
\description{
Given a Bin object with binStart and binEnd of class character,
convert both to numeric and return the Bin.
}
