% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Bin.R
\name{as.numeric,BinList-method}
\alias{as.numeric,BinList-method}
\title{Convert bins in a BinList with character start and end values to numeric}
\usage{
\S4method{as.numeric}{BinList}(x)
}
\arguments{
\item{x}{BinList}
}
\value{
BinList
}
\description{
Given a BinList object containing Bins that may have a character binStart or binEnd,
convert all binStarts and binEnds to numeric and return the BinList.
}
