% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Bin.R
\docType{class}
\name{Bin-class}
\alias{Bin-class}
\alias{Bin}
\title{Bin}
\description{
A class to define the start, end and label for a bin.
Optionally, may include a value for that bin.
If no start and end are provided, then assume the 'bin'
represents a category.
}
\section{Slots}{

\describe{
\item{\code{binStart}}{A number, Date or Posix value}

\item{\code{binEnd}}{A number, Date or Posix value}

\item{\code{binLabel}}{A string}

\item{\code{value}}{}
}}

