% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CollectionWithMetadata.R
\docType{class}
\name{CollectionWithMetadata-class}
\alias{CollectionWithMetadata-class}
\alias{CollectionWithMetadata}
\title{Collection With Metadata}
\description{
A class for working with numeric data 'collections' and associated metadata.
Variable collections should represent a biologically or scientifically coherent concept
measured over a consistent or comparable range of values. Examples include relative abundances of various genera,
or relative expression levels of genes in a gene expression profile.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame of collection values with variables as columns and records as rows}

\item{\code{sampleMetadata}}{A SampleMetadata object of metadata about the records where records are rows and metadata variables as columns}

\item{\code{recordIdColumn}}{The name of the column containing IDs for the records.}

\item{\code{ancestorIdColumns}}{A character vector of column names representing parent entities of the recordIdColumn.}

\item{\code{imputeZero}}{A logical indicating whether NA/ null values should be replaced with zeros.}

\item{\code{removeEmptySamples}}{A logical indicating whether empty (all NA/ zero) samples should be removed.}
}}

