% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ComputeResult.R
\docType{class}
\name{ComputeResult-class}
\alias{ComputeResult-class}
\alias{ComputeResult}
\title{Compute Result}
\description{
A class for consistently representing the results of computes.
This includes their representation in R, as JSON and how they are written to files.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame of values where computed variables are columns and samples rows.}

\item{\code{name}}{The name of the compute, ex: 'alphaDiv'.}

\item{\code{recordIdColumn}}{The name of the column containing IDs for the samples. All other columns will be treated as computed values.}

\item{\code{ancestorIdColumns}}{A character vector of column names representing parent entities of the recordIdColumn.}

\item{\code{computedVariableMetadata}}{mbioUtils::VariableMetadataList detailing the computed variables.}

\item{\code{statistics}}{An optional slot of any values. List or data.frame are recommended. It is not required to have rows or cols map to samples.}

\item{\code{computationDetails}}{An optional message about the computed results.}

\item{\code{parameters}}{A record of the input parameters used to generate the computed results.}

\item{\code{droppedColumns}}{A character vector of column names that have been dropped for being unsuitable for the computation.}
}}

