% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-correlation.R
\name{correlation}
\alias{correlation}
\alias{correlation,data.table,data.table-method}
\alias{correlation,data.table,missingOrNULL-method}
\alias{correlation,CollectionWithMetadata,missingOrNULL-method}
\alias{correlation,Collection,Collection-method}
\title{Correlation}
\usage{
correlation(
  data1,
  data2,
  method,
  format = c("ComputeResult", "data.table"),
  verbose = c(TRUE, FALSE),
  ...
)

\S4method{correlation}{data.table,data.table}(
  data1,
  data2,
  method = c("spearman", "pearson"),
  format = c("ComputeResult", "data.table"),
  verbose = c(TRUE, FALSE)
)

\S4method{correlation}{data.table,missingOrNULL}(
  data1,
  data2,
  method = c("spearman", "pearson", "sparcc"),
  format = c("ComputeResult", "data.table"),
  verbose = c(TRUE, FALSE)
)

\S4method{correlation}{CollectionWithMetadata,missingOrNULL}(
  data1,
  data2,
  method = c("spearman", "pearson"),
  format = c("ComputeResult", "data.table"),
  verbose = c(TRUE, FALSE),
  proportionNonZeroThreshold = 0.05,
  varianceThreshold = 0,
  stdDevThreshold = 0,
  metadataIsFirst = c(FALSE, TRUE)
)

\S4method{correlation}{Collection,Collection}(
  data1,
  data2,
  method = c("spearman", "pearson"),
  format = c("ComputeResult", "data.table"),
  verbose = c(TRUE, FALSE),
  proportionNonZeroThreshold = 0.05,
  varianceThreshold = 0,
  stdDevThreshold = 0
)
}
\arguments{
\item{data1}{first dataset. A data.table}

\item{data2}{second dataset. A data.table}

\item{method}{string defining the type of correlation to run.
The currently supported values are specific to the class of data1 and data2.}

\item{format}{string defining the desired format of the result.
The currently supported values are 'data.table' and 'ComputeResult'.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.frame with correlation coefficients or a ComputeResult object

data.frame with correlation coefficients
}
\description{
This function returns correlation coefficients for variables in one dataset against variables in a second dataset
}
