% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CountDataCollection.R
\docType{class}
\name{CountDataCollection-class}
\alias{CountDataCollection-class}
\alias{CountDataCollection}
\title{Count Data}
\description{
A class for working with count data, including microbial or genetic assays.
}
\section{Slots}{

\describe{
\item{\code{data}}{A data.frame of integer abundance counts with genes (species, etc.) as columns and samples as rows}

\item{\code{sampleMetadata}}{A data.frame of metadata about the samples with samples as rows and metadata variables as columns}

\item{\code{recordIdColumn}}{The name of the column containing IDs for the samples. All other columns will be treated as abundance values.}

\item{\code{ancestorIdColumns}}{A character vector of column names representing parent entities of the recordIdColumn.}

\item{\code{imputeZero}}{A logical indicating whether NA/ null values should be replaced with zeros.}
}}

