% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cut.R
\name{cut_interval}
\alias{cut_interval}
\title{Discretise numeric data into equal ranges}
\usage{
cut_interval(x, n = NULL, length = NULL, ...)
}
\arguments{
\item{x}{numeric vector}

\item{n}{number of intervals to create, OR}

\item{length}{length of each interval}

\item{...}{
  Arguments passed on to \code{\link[base:cut]{base::cut.default}}
  \describe{
    \item{\code{breaks}}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}
    \item{\code{labels}}{labels for the levels of the resulting category.  By default,
    labels are constructed using \code{"(a,b]"} interval notation.  If
    \code{labels = FALSE}, simple integer codes are returned instead of
    a factor.}
    \item{\code{right}}{logical, indicating if the intervals should be closed on
    the right (and open on the left) or vice versa.}
    \item{\code{dig.lab}}{integer which is used when labels are not given.  It
    determines the number of digits used in formatting the break numbers.}
    \item{\code{ordered_result}}{logical: should the result be an ordered factor?}
  }}
}
\description{
This function makes \code{n} groups with equal range
}
