% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{data_frame}
\alias{data_frame}
\title{Fast data.frame}
\usage{
data_frame(...)
}
\arguments{
\item{...}{usual arguments to data.frame}
}
\description{
This function should be used with extreme caution. If data.frame is not
performant enough, you should consider data.table instead. This is
intended to meet a very specific use case where a solution is needed which
is both efficient and tidy (meaning tidyverse compatible). It comes at the
cost of cutting some corners. There is no checking, recycling etc. unless asked for.
}
