% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-differentialExpression.R
\name{differentialExpression}
\alias{differentialExpression}
\title{Differential expression}
\usage{
differentialExpression(
  collection,
  comparator,
  method = c("DESeq", "limma"),
  pValueFloor = P_VALUE_FLOOR,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{collection}{CollectionWithMetadata object}

\item{comparator}{Comparator object specifying the variable and values or bins to be used in dividing samples into groups.}

\item{method}{string defining the differential expression method. Accepted values are 'DESeq' (for CountDataCollection) and 'limma' (for ArrayDataCollection).}

\item{pValueFloor}{numeric value that indicates the smallest p value that should be returned.
The corresponding adjusted p value floor will also be updated based on this value, and will be set to the maximum adjusted p value of all floored p values.
The default value uses the P_VALUE_FLOOR=1e-200 constant defined in this package.}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
ComputeResult object
}
\description{
This function returns the fold change and associated p value for a differential expression analysis comparing samples in two groups.
}
