% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{dropZeroColumns}
\alias{dropZeroColumns}
\title{Drop columns containing all 0s}
\usage{
dropZeroColumns(x, ...)
}
\arguments{
\item{x}{data.table, data.frame, or list}

\item{ignoreNA}{boolean indicating if NA values should be ignored when determining all unique values in a column.
Default is TRUE, which means a column that contains only 0 and NA values will be dropped. If FALSE, a column that contains 0s and
NAs will not be dropped.}
}
\value{
cleanedData. A list that includes 1. "data" the cleaned data, and 2. "droppedColumns" names of the
columns that have been removed.
}
\description{
This function drops all columns that have only 0 values.
}
