% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findAllColNames}
\alias{findAllColNames}
\title{All EDA Variable Column Names}
\usage{
findAllColNames(variables)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}
}
\value{
character vector of column names for all entries in the VariableMetadataList
}
\description{
This function provides EDA-compliant column names provided
an EDA-compliant VariableMetadataList object.
}
