% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findBinStart}
\alias{findBinStart}
\title{Find Bin Start}
\usage{
findBinStart(x, addTimeZone = c(FALSE, TRUE))
}
\arguments{
\item{x}{string that represents a bin. Ex: '[1, 2)'}

\item{addTimeZone}{boolean determines if time zone should be added to dates. Default is FALSE.}
}
\description{
findBinStart finds the start value of a given bin.
}
