% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findCollectionVariableMetadata}
\alias{findCollectionVariableMetadata}
\title{EDA Variable Metadata matching a Collection}
\usage{
findCollectionVariableMetadata(variables)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}
}
\value{
VariableMetadata object where isCollection is TRUE
}
\description{
This function returns a VariableMetadata object provided
an EDA-compliant VariableMetadataList object.
}
