% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findColNamesByPredicate}
\alias{findColNamesByPredicate}
\title{Find EDA column names}
\usage{
findColNamesByPredicate(variables, predicateFunction)
}
\arguments{
\item{variables}{a VariableMetadataList}

\item{predicateFunction}{a function to identify which VariableMetadata objects to return column names for}
}
\value{
character vector of column names for VariableMetadata objects matching the predicate
}
\description{
Find EDA column names
}
