% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findColNamesFromPlotRef}
\alias{findColNamesFromPlotRef}
\title{EDA Variable Column Names matching a PlotReference}
\usage{
findColNamesFromPlotRef(variables, plotRef)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}

\item{plotRef}{a string representing the PlotReference to look for}
}
\value{
character vector of column names matching the provided plot reference
}
\description{
This function provides EDA-compliant column names provided
an EDA-compliant VariableMetadataList object.
}
