% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findDataShapesFromPlotRef}
\alias{findDataShapesFromPlotRef}
\title{EDA Variable Data Shapes matching a PlotReference}
\usage{
findDataShapesFromPlotRef(variables, plotRef)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}

\item{plotRef}{a string representing the PlotReference to look for}
}
\value{
character vector of data shapes matching the provided plot reference
}
\description{
This function provides EDA-compliant data shapes provided
an EDA-compliant VariableMetadataList object.
}
