% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findDataTypesFromPlotRef}
\alias{findDataTypesFromPlotRef}
\title{EDA Variable Data Types matching a PlotReference}
\usage{
findDataTypesFromPlotRef(variables, plotRef)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}

\item{plotRef}{a string representing the PlotReference to look for}
}
\value{
character vector of data types matching the provided plot reference
}
\description{
This function provides EDA-compliant data types provided
an EDA-compliant VariableMetadataList object.
}
