% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findIndexFromPlotRef}
\alias{findIndexFromPlotRef}
\title{EDA Variable index matching a PlotReference}
\usage{
findIndexFromPlotRef(variables, plotRef)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}

\item{plotRef}{a string representing the PlotReference to look for}
}
\value{
numeric vector of indices matching the provided plot reference
}
\description{
This function returns a list index provided
an EDA-compliant VariableMetadataList object.
}
