% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{findNumericCols}
\alias{findNumericCols}
\title{Find numeric columns}
\usage{
findNumericCols(x)
}
\arguments{
\item{x}{list, data.frame, data.table, or vector}
}
\value{
vector of numeric column names in x. If no numeric columns found, returns NULL
}
\description{
This function finds all numeric columns in a data structure
}
