% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findStudyDependentVocabularyVariableMetadata}
\alias{findStudyDependentVocabularyVariableMetadata}
\title{EDA Variable Metadata with a Study-dependent Vocabulary}
\usage{
findStudyDependentVocabularyVariableMetadata(variables)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}
}
\value{
VariableMetadataList object where hasStudyDependentVocabulary is TRUE
}
\description{
This function returns a VariableMetadataList object provided
an EDA-compliant VariableMetadataList object. The resulting object
is a subset of the original and includes all elements where
\code{hasStudyDependentVocabulary} is TRUE.
}
