% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findVariablesNeedingWeightingVariableMetadata}
\alias{findVariablesNeedingWeightingVariableMetadata}
\title{EDA Variable Metadata which needs weighting}
\usage{
findVariablesNeedingWeightingVariableMetadata(variables)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}
}
\value{
VariableMetadataList object where weightingVariableSpec is not NULL/ NA
}
\description{
This function returns a VariableMetadataList object provided
an EDA-compliant VariableMetadataList object. The resulting object
is a subset of the original and includes all elements where
\code{weightingVariableSpec} is not NULL/ NA.
}
