% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findVariableSpecFromPlotRef}
\alias{findVariableSpecFromPlotRef}
\title{EDA Variable Spec matching a PlotReference}
\usage{
findVariableSpecFromPlotRef(variables, plotRef)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}

\item{plotRef}{a string representing the PlotReference to look for}
}
\value{
VariableMetadata object matching the provided plot reference
}
\description{
This function returns a VariableSpec object provided
an EDA-compliant VariableMetadataList object.
}
