% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{findWeightingVariablesMetadata}
\alias{findWeightingVariablesMetadata}
\title{EDA VariableMetadataList matching any weightingVariableSpec}
\usage{
findWeightingVariablesMetadata(variables)
}
\arguments{
\item{variables}{a VariableMetadataList of variables to search}
}
\value{
VariableMetadataList object where variableSpec matches any elements specified weightingVariableSpec
}
\description{
This function returns a VariableMetadataList object provided
an EDA-compliant VariableMetadataList object. The resulting object
will be a subset of the original and include only those elements
for which another variable has specified it as an annotated
\code{weightingVariableSpec}.
}
