% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAggStr}
\alias{getAggStr}
\title{Make Aggregation Formulas}
\usage{
getAggStr(LHS = NULL, RHS = NULL)
}
\arguments{
\item{LHS}{vector of variable/ column names to become the LHS}

\item{RHS}{vector of variable/ column names to become the RHS}
}
\value{
character string which can be used as input to \code{aggregate} with \code{as.formula}
}
\description{
This function is intended to help create formulas for input to
base R's \code{aggregate} function.
}
