% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Collections.R
\name{getCollectionNames}
\alias{getCollectionNames}
\alias{getCollectionNames,Collections-method}
\title{Get Names of Collections}
\usage{
getCollectionNames(object)

\S4method{getCollectionNames}{Collections}(object)
}
\arguments{
\item{object}{A Collections object, or MbioDataset}
}
\value{
A character vector of collection names
}
\description{
Get the names of the collections in the Collections or MbioDataset object
}
