% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CollectionWithMetadata.R
\name{getCollectionValues}
\alias{getCollectionValues}
\alias{getCollectionValues,CollectionWithMetadata-method}
\title{Get data.table of values from CollectionWithMetadata}
\usage{
getCollectionValues(
  object,
  ignoreImputeZero = c(FALSE, TRUE),
  includeIds = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE)
)

\S4method{getCollectionValues}{CollectionWithMetadata}(
  object,
  ignoreImputeZero = c(FALSE, TRUE),
  includeIds = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{object}{CollectionWithMetadata}

\item{ignoreImputeZero}{boolean indicating whether we should respect the imputeZero slot}

\item{includeIds}{boolean indicating whether we should include recordIdColumn and ancestorIdColumns}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table of values
}
\description{
Returns a data.table of collection values, respecting the
\code{imputeZero} slot.
}
