% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Collections.R
\name{getCollectionVariableNames}
\alias{getCollectionVariableNames}
\alias{getCollectionVariableNames,Collection-method}
\title{Get Collection Variable Names}
\usage{
getCollectionVariableNames(object, ...)

\S4method{getCollectionVariableNames}{Collection}(object)
}
\arguments{
\item{object}{A Collection}

\item{...}{Additional arguments}
}
\value{
A character vector of variable names
}
\description{
Get the names of the variables in the Collection
}
