% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{getColName}
\alias{getColName}
\title{EDA Variable Column Name of a VariableSpec}
\usage{
getColName(varSpec)
}
\arguments{
\item{varSpec}{a VariableSpec to find the column name for}
}
\value{
a character vector of length one
}
\description{
This function provides an EDA-compliant column name for a
particular VariableSpec.
}
