% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Statistic.R
\name{getDataTable}
\alias{getDataTable}
\title{Convert StatisicList to data.table}
\usage{
getDataTable(object)
}
\arguments{
\item{object}{A StatisticList object}
}
\value{
data.table
}
\description{
This function returns a data.table representation of a StatisticList.
The content of the slots are stored as JSON. Each entry in the list becomes a named column.
}
