% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{getDiscretizedBins}
\alias{getDiscretizedBins}
\title{Find Bin Ranges for a Continuous Variable}
\usage{
getDiscretizedBins(
  x,
  method = c("equalInterval", "quantile", "sd"),
  numBins = NULL,
  getValue = c(TRUE, FALSE)
)
}
\arguments{
\item{x}{Numeric (or Date) vector to find bins for}

\item{method}{A string indicating which method to use to find bins ('equalRanges', 'quantile', 'sd')}

\item{numBins}{A number indicating how many bins are desired}

\item{getValue}{A boolean indicating whether to return the counts per bin}
}
\description{
This function will find bin start, end and labels for a
continuous variable. Optionally, it can return a value/ count
per bin. By default returns 10 bins for equalRanges and quantile
methods and 6 for sd (standard deviations).
}
