% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Megastudy.R
\name{getDTWithImputedZeroes}
\alias{getDTWithImputedZeroes}
\title{Impute Zeroes (on tall data)}
\usage{
getDTWithImputedZeroes(object, variables)
}
\arguments{
\item{object}{veupathUtils::Megastudy}
}
\value{
data.table
}
\description{
This function returns a data.table which has explicit zero values
for all expected categories of some variable of interest in a megastudy.
}
