% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CollectionWithMetadata.R
\name{getMetadataVariableNames}
\alias{getMetadataVariableNames}
\alias{getMetadataVariableNames,CollectionWithMetadata-method}
\title{Get Variable Names of Metadata}
\usage{
getMetadataVariableNames(object)

\S4method{getMetadataVariableNames}{CollectionWithMetadata}(object)
}
\arguments{
\item{object}{Any object with a slot containing sampleMetadata}
}
\value{
a character vector of metadata variable names
}
\description{
Get the names of the metadata variables in an object containing sample metadata.
}
