% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CollectionWithMetadata.R
\name{getMetadataVariableSummary}
\alias{getMetadataVariableSummary}
\alias{getMetadataVariableSummary,CollectionWithMetadata-method}
\title{Get Summary of Metadata Variables}
\usage{
getMetadataVariableSummary(object, variable)

\S4method{getMetadataVariableSummary}{CollectionWithMetadata}(object, variable)
}
\arguments{
\item{object}{An object with a slot containing sampleMetadata}

\item{variable}{A character vector representing the name of the metadata variable to summarize}
}
\value{
a table summarizing the values of the requested metadata variable
}
\description{
Get a summary of the requested metadata variable in an object containing sample metadata.
}
