% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CollectionWithMetadata.R,
%   R/methods-SampleMetadata.R
\name{getSampleMetadata}
\alias{getSampleMetadata}
\alias{getSampleMetadata,CollectionWithMetadata-method}
\alias{getSampleMetadata,SampleMetadata-method}
\title{Get data.table of sample metadata from CollectionWithMetadata}
\usage{
getSampleMetadata(
  object,
  asCopy = c(TRUE, FALSE),
  includeIds = c(TRUE, FALSE),
  metadataVariables = NULL
)

\S4method{getSampleMetadata}{CollectionWithMetadata}(
  object,
  asCopy = c(TRUE, FALSE),
  includeIds = c(TRUE, FALSE),
  metadataVariables = NULL
)

\S4method{getSampleMetadata}{SampleMetadata}(object, asCopy = c(TRUE, FALSE), includeIds = c(TRUE, FALSE))
}
\arguments{
\item{object}{CollectionWithMetadata}

\item{asCopy}{boolean indicating whether to return the data as a copy or by reference}

\item{includeIds}{boolean indicating whether we should include recordIdColumn and ancestorIdColumns}

\item{metadataVariables}{The metadata variables to include in the sample metadata. If NULL, all metadata variables will be included.}
}
\value{
data.table of sample metadata
}
\description{
Returns a data.table of sample metadata
}
