% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CollectionWithMetadata.R,
%   R/methods-SampleMetadata.R
\name{getSampleMetadataIdColumns}
\alias{getSampleMetadataIdColumns}
\alias{getSampleMetadataIdColumns,CollectionWithMetadata-method}
\alias{getSampleMetadataIdColumns,SampleMetadata-method}
\title{Get Sample Metadata Id Column Names}
\usage{
getSampleMetadataIdColumns(object)

\S4method{getSampleMetadataIdColumns}{CollectionWithMetadata}(object)

\S4method{getSampleMetadataIdColumns}{SampleMetadata}(object)
}
\arguments{
\item{object}{A Microbiome Dataset, or other object w sample metadata}
}
\value{
a character vector of id column names
}
\description{
Get the names of the record and ancestor id columns in the sample metadata of the Microbiome Dataset.
}
