% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Megastudy.R
\name{getStudyIdColumnName}
\alias{getStudyIdColumnName}
\title{StuydIdColName as String}
\usage{
getStudyIdColumnName(object)
}
\arguments{
\item{object}{veupathUtils::StudySpecificVocabulariesByVariable}
}
\value{
character
}
\description{
This function returns the studyIdColName from an StudySpecificVocabulariesByVariable
}
