% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Megastudy.R
\name{getVariableSpec}
\alias{getVariableSpec}
\title{get a VariableSpec}
\usage{
getVariableSpec(object, ...)
}
\arguments{
\item{object}{An object containing a veupathUtils::VariableSpec}
}
\value{
character
}
\description{
This function returns a string representation of a VariableSpec. By
default it assumes the VariableSpec is available in a slot called
\code{variableSpec}.
}
