% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Megastudy.R
\name{getVariableSpecColumnName}
\alias{getVariableSpecColumnName}
\title{VarSpecColName as String}
\usage{
getVariableSpecColumnName(object)
}
\arguments{
\item{object}{veupathUtils::StudySpecificVocabulariesByVariable}
}
\value{
character
}
\description{
This function returns the variableSpec from an StudySpecificVocabulariesByVariable
}
