% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Collections.R
\name{isOneToManyWithAncestor}
\alias{isOneToManyWithAncestor}
\alias{isOneToManyWithAncestor,Collection-method}
\title{Is One to Many With Ancestor}
\usage{
isOneToManyWithAncestor(object, verbose = c(TRUE, FALSE))

\S4method{isOneToManyWithAncestor}{Collection}(object, verbose = c(TRUE, FALSE))
}
\arguments{
\item{object}{Collection}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
boolean
}
\description{
Determines if the collection is one-to-many with its ancestor(s).
Importantly, if there are no ancestors, this function returns FALSE.
}
