% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbioUtils-package.R
\docType{package}
\name{mbioUtils-package}
\alias{mbioUtils}
\alias{mbioUtils-package}
\title{mbioUtils: Utility Functions for MicrobiomeDB R Packages}
\description{
mbioUtils contains utility functions and data structures for microbiomeDB R packages, including microbiomeComputations, MicrobiomeDB, and microbiomeData. Forked from veupathUtils 2.6.7.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/microbiomeDB/mbioUtils}
  \item Report bugs at \url{https://github.com/microbiomeDB/mbioUtils/issues}
}

}
\author{
\strong{Maintainer}: Danielle Callan \email{dcallan@upenn.edu}

Authors:
\itemize{
  \item Ann Blevins \email{annsize@upenn.edu} (Original veupathUtils author)
}

}
\keyword{internal}
