% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{merge}
\alias{merge}
\title{Merge Variable*List objects}
\usage{
merge(x, y, ...)
}
\arguments{
\item{x}{Variable*List object}

\item{y}{Variable*List object}
}
\value{
Variable*List object containing the entries from both inputs
}
\description{
This function takes two VariableMetadataList or VariableSpecList objects and returns one.
}
