% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{naToZero}
\alias{naToZero}
\title{Replace numeric NAs with 0}
\usage{
naToZero(x, ...)
}
\arguments{
\item{x}{list, data.frame, array, or vector}

\item{cols}{Optional. When appropriate, vector of column names
for which the NA replacement should occur.
Default is all numeric columns.}
}
\value{
x with desired NAs replaced with 0.
}
\description{
This function replaces numeric NAs with 0.
}
