% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nonZeroRound}
\alias{nonZeroRound}
\title{Non-Zero Rounding}
\usage{
nonZeroRound(x, digits)
}
\arguments{
\item{x}{Numeric value to round}

\item{digits}{Number indicating the desired precision}
}
\value{
number rounded as nearly to the requested precision as
possible without returning zero.
}
\description{
This function will recursively attempt to round a value to
greater and greater precision until it results in a non-zero
value. One consequence of this is that the precision of the
output value may not be exactly what was requested.
}
