% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-pca.R
\name{pca}
\alias{pca}
\title{PCA}
\usage{
pca(collection, nPCs = 10, ntop = 500, verbose = c(TRUE, FALSE))
}
\arguments{
\item{collection}{A Collection object}

\item{nPCs}{Number of principal components to return. Default 10}

\item{ntop}{Use the top ntop genes with the highest variance for the pca computation. Mirrors the deseq2 plotPCA argument. Default 500.}

\item{verbose}{Boolean indicating if extra messaging should be printed.}
}
\value{
A data table with the id columns and the first nPCs principal components.
}
\description{
PCA
}
