% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Collections.R
\name{pruneFeatures}
\alias{pruneFeatures}
\alias{pruneFeatures,Collection-method}
\title{Prune features by predicate}
\usage{
pruneFeatures(object, predicate, verbose = c(TRUE, FALSE))

\S4method{pruneFeatures}{Collection}(object, predicate, verbose = c(TRUE, FALSE))
}
\arguments{
\item{object}{Collection}

\item{predicate}{Function returning a boolean indicating if a feature should be included (TRUE) or excluded (FALSE)}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
Collection with modified data slot
}
\description{
Modifies the data slot of a Collection or
CollectionWithMetadata object, to exclude features for which
the provided predicate function returns FALSE.
}
