% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Range.R
\docType{class}
\name{Range-class}
\alias{Range-class}
\alias{Range}
\title{Data Range}
\description{
A class to a data range, represented as minimum and maximum values.
The range could be numeric or Date/ Posix
}
\section{Slots}{

\describe{
\item{\code{minimum}}{A number, Date or Posix value}

\item{\code{maximum}}{A number, Date or Posix value}
}}

