% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CollectionWithMetadata.R
\name{removeIncompleteRecords}
\alias{removeIncompleteRecords}
\alias{removeIncompleteRecords,CollectionWithMetadata-method}
\title{Drop Records with incomplete SampleMetadata}
\usage{
removeIncompleteRecords(
  object,
  colName = character(),
  verbose = c(TRUE, FALSE)
)

\S4method{removeIncompleteRecords}{CollectionWithMetadata}(
  object,
  colName = character(),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{object}{CollectionWithMetadata}

\item{colName}{String providing the column name in SampleMetadata to check for completeness}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
CollectionWithMetadata with modified data and sampleMetadata slots
}
\description{
Modifies the data and sampleMetadata slots of an
CollectionWithMetadata object, to exclude records with
missing SampleMetadata for a specified column.
}
