% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescaleToNonNeg}
\alias{rescaleToNonNeg}
\title{Rescale Values to be non-negative}
\usage{
rescaleToNonNeg(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function rescales all values in a vector to be non-negative.
}
