% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classes.R
\name{setAttrFromList}
\alias{setAttrFromList}
\title{Set object attributes from a list}
\usage{
setAttrFromList(.dt, attr, removeExtraAttrs = T)
}
\arguments{
\item{.dt}{data.table}

\item{attr}{named list of desired attributes to add to .dt}

\item{removeExtraAttrs}{boolean indicating if existing attributes of .dt not
included in attr should be removed.}
}
\description{
This function sets the attributes of an object based on a given list
}
