% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{setNaToValue}
\alias{setNaToValue}
\title{Impute 0s for numeric columns}
\usage{
setNaToValue(dt, value = 0, cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{value}{Numeric value to use when replacing NAs. Default is 0.}

\item{cols}{Vector of column names for which the NA replacement should occur.
Defulat is all numeric columns.}
}
\value{
inputted data.table dt but with NAs replaced with desired value.
}
\description{
This function replaces NAs in numeric columns of a data.table
with a desired value.
}
