% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{setNaToZero}
\alias{setNaToZero}
\title{Impute 0s for numeric columns}
\usage{
setNaToZero(dt, cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{cols}{Vector of column names for which the NA replacement should occur.
Default is all numeric columns.}
}
\value{
inputted data.table dt but with NAs replaced with desired value.
}
\description{
This function replaces NAs in numeric columns of a data.table with 0.
}
