% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setroworder}
\alias{setroworder}
\title{Set Row Order}
\usage{
setroworder(x, neworder)
}
\arguments{
\item{neworder}{Character vector of the new column name ordering.}

\item{a}{data.table}
}
\description{
setroworder reorders the rows of a data.table to the new order provided.
}
