% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shiftToNonNeg}
\alias{shiftToNonNeg}
\title{Shift Values to be non-negative}
\usage{
shiftToNonNeg(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function shifts all values in a vector to be non-negative.
}
