% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Statistic.R
\docType{class}
\name{Statistic-class}
\alias{Statistic-class}
\alias{Statistic}
\title{Statistic}
\description{
A class to specify a named statistic, its value, confidence interval and p-value.
This is primarily to help maintain consistency in meeting the EDA API.
}
\section{Slots}{

\describe{
\item{\code{name}}{A string specifying what statistic was calculated}

\item{\code{value}}{A number}

\item{\code{confidenceInterval}}{An optional Range specifying minimum and maximum values for the confidence interval}

\item{\code{confidenceLevel}}{An optional decimal number indicating the degree of confidence represented by the confidence interval}

\item{\code{pvalue}}{An optional string representing the p-value associated with the statistic. The string should either be coercible to a number or be '<0.0001'.}
}}

