% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stripEntityIdFromColumnHeader}
\alias{stripEntityIdFromColumnHeader}
\title{Strip Entity ID from Column Header}
\usage{
stripEntityIdFromColumnHeader(columnNames)
}
\arguments{
\item{columnNames}{character vector}
}
\description{
This function strips entity ID from column headers for EDA formatted tabular data.
}
