% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testArrayDataCollection}
\alias{testArrayDataCollection}
\title{Test Array Data Collection}
\format{
\subsection{\code{testArrayDataCollection}}{

An ArrayDataCollection object with:
\describe{
\item{data}{50 samples x 100 antibody targets (continuous expression values)}
\item{sampleMetadata}{Includes treatment, batch, age, and sex variables}
\item{recordIdColumn}{"entity.SampleID"}
}

Sample metadata variables:
\describe{
\item{entity.treatment}{Binary treatment variable (treatment_A, treatment_B)}
\item{entity.batch}{Batch variable (batch_1 through batch_5)}
\item{entity.age}{Continuous age variable (normal distribution, mean 45, sd 10)}
\item{entity.sex}{Binary sex variable (male, female)}
}
}
}
\usage{
testArrayDataCollection
}
\description{
An ArrayDataCollection object created from testArrayData with associated sample metadata.
This object is used for testing the ArrayDataCollection class and limma-based differential
expression methods.
}
\keyword{datasets}
