% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toColNameOrNull}
\alias{toColNameOrNull}
\title{EDA Variable Column Names}
\usage{
toColNameOrNull(varDetailsList)
}
\arguments{
\item{varDetailsList}{a named list containing at minimum 'variableId', 'entityId'}
}
\value{
character string representing the expected column name for that variable
}
\description{
This function provides EDA-compliant column names provided
EDA-compliant VariableDetails.
}
