% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-string.R
\name{toStringOrNull}
\alias{toStringOrNull}
\title{Replace Empty String with NULL}
\usage{
toStringOrNull(x = NULL, na.rm = c(TRUE, FALSE))
}
\arguments{
\item{x}{character vector}
}
\value{
non-empty character vector or NULL
}
\description{
This function replaces the empty string "" with NULL
}
