% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim}
\alias{trim}
\title{Trim White Space}
\usage{
trim(x)
}
\arguments{
\item{x}{Character vector to trim white space from.}
}
\value{
character vector of trimmed values
}
\description{
Function to trim white space from a string.
}
