% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-VariableMetadata.R
\docType{class}
\name{VariableSpec-class}
\alias{VariableSpec-class}
\alias{VariableSpec}
\title{Variable Specification}
\description{
A class for working with variable specifications. Contains a variable ID and an entity ID.
}
\section{Slots}{

\describe{
\item{\code{variableId}}{The name of the variable}

\item{\code{entityId}}{The name of the entity}
}}

