% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veupathUtils-package.R
\docType{package}
\name{veupathUtils-package}
\alias{veupathUtils}
\alias{veupathUtils-package}
\title{veupathUtils: General Helper Functions for VEuPathDB Projects}
\description{
veupathUtils contains various R helper functions intended to be useful across a variety of projects, including the plot.data and microbiomeComputations packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VEuPathDB/veupathUtils}
  \item Report bugs at \url{https://github.com/VEuPathDB/veupathUtils/issues}
}

}
\author{
\strong{Maintainer}: Ann Blevins \email{annsize@upenn.edu}

Authors:
\itemize{
  \item Danielle Callan \email{dcallan@upenn.edu}
}

}
\keyword{internal}
