% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whichInBin}
\alias{whichInBin}
\title{Which values are In the Bin?}
\usage{
whichInBin(values, binString)
}
\arguments{
\item{values}{numeric vector}

\item{binString}{a string that defines a numeric bin. For example '[1, 2)'}
}
\description{
whichInBin finds array values that fall within a given bin
}
