% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Bin.R
\name{whichValuesInBin}
\alias{whichValuesInBin}
\title{Find values that fall within a Bin}
\usage{
whichValuesInBin(values, bin)
}
\arguments{
\item{values}{numeric vector}

\item{bin}{A Bin object}
}
\value{
logical vector
}
\description{
Given a numeric vector and a Bin, returns a logical vector indicating if
the value in the input vector falls within the bin.
}
