% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Bin.R
\name{whichValuesInBinList}
\alias{whichValuesInBinList}
\title{Find values in a vector that fall within any bins in a BinList}
\usage{
whichValuesInBinList(values, binList)
}
\arguments{
\item{values}{numeric vector}

\item{binList}{A BinList object}
}
\value{
logical vector
}
\description{
Given a vector of values and a BinList, return a logical vector that
indicates if a value was found in any Bin within the BinList
}
