% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{findNumericCols}
\alias{findNumericCols}
\title{Find numeric columns}
\usage{
findNumericCols(x)
}
\arguments{
\item{x}{list, data.frame, array, or vector}
}
\value{
vector of numeric column names in x
}
\description{
This function finds all numeric columns in a list
}
