% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Collections.R
\name{getCollectionData}
\alias{getCollectionData}
\alias{getCollectionData,Collection-method}
\title{Get data.table of values from Collection}
\usage{
getCollectionData(
  object,
  variableNames = NULL,
  ignoreImputeZero = c(FALSE, TRUE),
  includeIds = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE)
)

\S4method{getCollectionData}{Collection}(
  object,
  variableNames = NULL,
  ignoreImputeZero = c(FALSE, TRUE),
  includeIds = c(TRUE, FALSE),
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{object}{Collection}

\item{variableNames}{A character vector representing the name of the variables to return.
If NULL, returns all variables}

\item{ignoreImputeZero}{boolean indicating whether we should respect the imputeZero slot}

\item{includeIds}{boolean indicating whether we should include recordIdColumn and ancestorIdColumns}

\item{verbose}{boolean indicating if timed logging is desired}
}
\value{
data.table of values
}
\description{
Returns a data.table of collection values, respecting the
\code{imputeZero} slot.
}
