% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{setNaToZero}
\alias{setNaToZero}
\title{Replace numeric NAs with 0 - update by reference}
\usage{
setNaToZero(df, cols = NULL)
}
\arguments{
\item{cols}{vector of column names for which the NA replacement should occur.
Default is all numeric columns.}

\item{x}{data.table, data.frame, or list}
}
\value{
x with desired NAs replaced with 0.
}
\description{
This function replaces NAs in numeric columns with 0.
}
