% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Megastudy.R
\docType{class}
\name{StudySpecificVocabulariesByVariable-class}
\alias{StudySpecificVocabulariesByVariable-class}
\alias{StudySpecificVocabulariesByVariable}
\title{Study Specific Vocabularies By Variable}
\description{
A class to specify expected values per study for some variable
of interest.
}
