% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-VariableMetadata.R
\name{toJSON}
\alias{toJSON}
\title{R object as JSON string}
\usage{
toJSON(object, named = c(TRUE, FALSE))
}
\arguments{
\item{object}{object of a supported S4 class to convert to a JSON string representation}

\item{named}{logical indicating whether the result should be a complete named JSON object or just the value of the object}
}
\value{
character vector of length 1 containing JSON string
}
\description{
This function converts an R object to a JSON string.
see \code{methods(veupathUtils::toJSON)} for a list of support classes.
}
