% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-numeric.R
\name{validateNumericCols}
\alias{validateNumericCols}
\title{Validate numeric columns}
\usage{
validateNumericCols(x, cols, ...)
}
\arguments{
\item{x}{list, data.frame, array, or vector}

\item{cols}{vector of column names}
}
\value{
given column names
}
\description{
Given a vector of column names, this function ensures all
referenced columns are numeric
}
